+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+v3.6.2
+------
+
+- Fix incorrect usage of `config.get` left over from previous fixes
+
+
+v3.6.1
+------
+
+- Fix incorrect usage of `get_config`
+
+
 v3.6.0
 ------